move to Enum namespace
