+++ lib/DTO/NodeConfiguration.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Einenlum\PhpStackDetector\DTO;
 
-use Einenlum\PhpStackDetector\NodePackageManagerType;
+use Einenlum\PhpStackDetector\Enum\NodePackageManagerType;
 
 readonly class NodeConfiguration
 {
