+++ lib/StackDetector/Typo3Detector.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Einenlum\PhpStackDetector\StackDetector;
 
 use Einenlum\PhpStackDetector\StackDetectorInterface;
-use Einenlum\PhpStackDetector\StackType;
+use Einenlum\PhpStackDetector\Enum\StackType;
 
 class Typo3Detector extends BaseComposerTypeDetector implements StackDetectorInterface
 {
