+++ lib/StackDetector/WordpressDetector.php
@@ -6,7 +6,7 @@ use Einenlum\PhpStackDetector\DirectoryCrawler\AdapterInterface;
 use Einenlum\PhpStackDetector\Exception\ResourceNotFoundException;
 use Einenlum\PhpStackDetector\DTO\Stack;
 use Einenlum\PhpStackDetector\StackDetectorInterface;
-use Einenlum\PhpStackDetector\StackType;
+use Einenlum\PhpStackDetector\Enum\StackType;
 
 class WordpressDetector implements StackDetectorInterface
 {
